--- libs/cdk/internals/rxjs/src/index.ts
@@ -1,4 +1,0 @@
-export { immediateProvider } from './lib/scheduler/immediateProvider';
-export { animationFrameProvider } from './lib/scheduler/animationFrameProvider';
-export { intervalProvider } from './lib/scheduler/intervalProvider';
-export { dateTimestampProvider } from './lib/scheduler/date-time-stamp.provider';
