--- libs/cdk/testing/README.md
@@ -1,29 +1,0 @@
-# @rx-angular/cdk/testing
-
-[![npm](https://img.shields.io/npm/v/%40rx-angular%2Fcdk.svg)](https://www.npmjs.com/package/%40rx-angular%2Fcdk)
-![rx-angular CI](https://github.com/rx-angular/rx-angular/workflows/rx-angular%20CI/badge.svg?branch=main)
-[![Coverage Status](https://raw.githubusercontent.com/rx-angular/rx-angular/github-pages/docs/test-coverage/cdk/jest-coverage-badge.svg)](https://rx-angular.github.io/rx-angular/test-coverage/cdk/lcov-report/index.html)
-
-## Testing utilities for `@rx-angular/cdk`
-
-`@rx-angular/cdk/testing` ... TDB
-
-## Key features
-
-- ✅ TestScheduler
-
-## Demos:
-
-- [⚡ Demo]()
-
-## Install
-
-```bash
-npm install --save @rx-angular/cdk
-# or
-yarn add @rx-angular/cdk
-```
-
-## Documentation
-
-- [Docs](./docs/Readme.md)
