--- libs/cdk/testing/docs/Readme.md
@@ -1,1 +1,0 @@
-# Testing Utilities for `@rx-angular/cdk`
