--- libs/cdk/testing/ng-package.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "../../../node_modules/ng-packagr/package.schema.json",
-  "lib": {
-    "entryFile": "src/index.ts",
-    "flatModuleFile": "cdk-testing"
-  }
-}
