--- libs/cdk/testing/src/index.ts
@@ -1,1 +1,0 @@
-export { RxTestScheduler } from './lib/test-scheduler';
