+++ libs/cdk/zone-less/rxjs/src/internals/date-time-stamp.provider.ts
