+++ libs/cdk/zone-less/rxjs/src/observable/fromEvent.ts
@@ -64,6 +64,7 @@ export interface AddEventListenerOptions extends EventListenerOptions {
   passive?: boolean;
 }
 
+/** @deprecated will be removed with 1.0.0 */
 export function fromEvent<T>(
   target: FromEventTarget<T>,
   eventName: string
@@ -74,6 +75,7 @@ export function fromEvent<T>(
   eventName: string,
   resultSelector: (...args: any[]) => T
 ): Observable<T>;
+/** @deprecated will be removed with 1.0.0 */
 export function fromEvent<T>(
   target: FromEventTarget<T>,
   eventName: string,
@@ -86,7 +88,7 @@ export function fromEvent<T>(
   options: EventListenerOptions,
   resultSelector: (...args: any[]) => T
 ): Observable<T>;
-
+/** @deprecated will be removed with 1.0.0 */
 export function fromEvent<T>(
   target: FromEventTarget<T>,
   eventName: string,
