+++ libs/cdk/zone-less/rxjs/src/scheduler/animation-frame/animationFrame.ts
@@ -33,6 +33,7 @@ import { AnimationFrameScheduler } from './AnimationFrameScheduler';
  *
  * // You will see a div element growing in height
  * ```
+ * @deprecated will be removed with 1.0.0
  */
 export const animationFrameScheduler = new AnimationFrameScheduler(
   AnimationFrameAction
